+++ src/modules/texteditor/index.js
@@ -1,4 +1,10 @@
 module.exports = (function() {
   'use strict';
 
+  angular.module('form-texteditor', [
+    'form-decorators',
+    'pascalprecht.translate',
+    'schemaForm',
+    'summernote'
+  ]).config(require('./form-texteditor'));
 })();
